+++ scripts/module/constants.js
@@ -5,6 +5,13 @@ const CONSTANTS = {
     PRIMARY: 0xEE442F,
     SECONDARY: 0x601A4A,
     TERTIARY: 0x601A4A
+  },
+  SHAPES: {
+    POLY: "polygon",
+    RECT: "rectangle",
+    CIRC: "circle",
+    ELIP: "ellipse",
+    RREC: "roundedRect",
   }
 }